Split feedforwardNN.py into multiple python files, reduce output, fix fold=0 vs fold=1
